requirements.txt
fastapi
uvicorn
face-recognition
pillow
psycopg2-binary
numpy
python-multipart
